Move transaction delete to transaction detail
